x86: boot: Fix EFI stub alignment
Forwarded: https://lore.kernel.org/lkml/20180916152246.GG4765@decadent.org.uk/T/#u

We currently align the end of the compressed image to a multiple of
16.  However the PE-COFF header included in the EFI stub says that the
file alignment is 32 bytes, and when adding an EFI signature to the
file it must first be padded to this alignment.

sbsigntool commands warn about this:

    warning: file-aligned section .text extends beyond end of file
    warning: checksum areas are greater than image size. Invalid section table?

Worse, pesign-at least when creating a detached signature—uses the
hash of the unpadded file, resulting in an invalid signature if
padding is required.

Avoid both these problems by increasing alignment to 32 bytes when
CONFIG_EFI_STUB is enabled.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>

Gbp-Pq: Topic bugfix/x86
Gbp-Pq: Name x86-boot-fix-efi-stub-alignment.patch